+++ src/deploy-output.ts
@@ -1,26 +1,12 @@
 import * as core from '@actions/core'
-import { getAccessToken } from './utils'
-import Client from 'platformsh-client'
+import { getCliClient, getEnvironmentName } from './utils'
 
 export async function deployOutput(): Promise<void> {
   core.startGroup('Output deployed URL')
 
-  let envName = core.getInput('environment-name')
-  if (!envName) {
-    const { GITHUB_REF_NAME } = process.env
-    if (GITHUB_REF_NAME) {
-      envName = GITHUB_REF_NAME
-    }
-  }
+  const client = await getCliClient(core.getInput('cli-token'))
+  const envName = getEnvironmentName()
 
-  const accessToken = await getAccessToken(core.getInput('cli-token'))
-  const client = new Client({
-    access_token: accessToken,
-    api_url: 'https://api.platform.sh/api',
-    authorization: ''
-  })
-
-  // Get env details
   const envResult = await client.getEnvironment(
     core.getInput('project-id'),
     encodeURIComponent(envName)
@@ -34,3 +20,4 @@ export async function deployOutput(): Promise<void> {
   core.setOutput('deployed-url', url)
   core.endGroup()
 }
+
